refactor(codegen): extract parts into plugin folder
