+++ scripts/codegen/src/core/registry.ts
@@ -12,6 +12,7 @@ import { htmlAttrsPlugin } from "../plugins/htmlAttrs/index.ts";
 import { imperativePropsPlugin } from "../plugins/imperativeProps/index.ts";
 import { latchPlugin } from "../plugins/latch/index.ts";
 import { motionPlugin } from "../plugins/motion/index.ts";
+import { partsPlugin } from "../plugins/parts/index.ts";
 import { polymorphicPlugin } from "../plugins/polymorphic/index.ts";
 import { propsPlugin } from "../plugins/props/index.ts";
 import { rootElementStaticPlugin } from "../plugins/rootElementStatic/index.ts";
@@ -48,4 +49,5 @@ export const PLUGINS: readonly SubstratePlugin[] = [
   eventsPlugin,
   vocabularyPlugin,
   constraintsPlugin,
+  partsPlugin,
 ];
